requirements.txt
attrs==23.2.0
Automat==22.10.0
balena-sdk @ git+https://github.com/balena-io/balena-sdk-python.git@c6fa01898e7d46bfd84915ede9d51b6a87625ad6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
Deprecated==1.2.14
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pine-client==0.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
ratelimit==2.2.1
requests==2.31.0
semver==3.0.2
service-identity==24.1.0
six==1.16.0
Twisted==24.3.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
zope.interface==6.3
